+++ modules/dxp/apps/commerce/commerce-api/src/main/java/com/liferay/commerce/price/CommercePriceFormatter.java
@@ -19,6 +19,8 @@ import aQute.bnd.annotation.ProviderType;
 import com.liferay.commerce.currency.model.CommerceCurrency;
 import com.liferay.portal.kernel.exception.PortalException;
 
+import java.math.BigDecimal;
+
 /**
  * @author Marco Leo
  * @author Alessio Antonio Rendina
@@ -26,8 +28,8 @@ import com.liferay.portal.kernel.exception.PortalException;
 @ProviderType
 public interface CommercePriceFormatter {
 
-	public String format(CommerceCurrency commerceCurrency, double price);
+	public String format(CommerceCurrency commerceCurrency, BigDecimal price);
 
-	public String format(long groupId, double price) throws PortalException;
+	public String format(long groupId, BigDecimal price) throws PortalException;
 
 }
